+++ src/Interface/ArrowButton.cpp
@@ -123,7 +123,7 @@ void ArrowButton::draw()
 	lock();
 
 	SDL_Rect rect; // draw button
-	Uint8 color (static_cast<Uint8>(static_cast<unsigned>(_color) + 2u));
+	Uint8 color (static_cast<Uint8>(_color + 2u));
 
 	rect.x =
 	rect.y = 0;
@@ -134,29 +134,29 @@ void ArrowButton::draw()
 
 	++rect.x;
 	++rect.y;
-	color = static_cast<Uint8>(static_cast<unsigned>(_color) + 5u);
+	color = static_cast<Uint8>(_color + 5u);
 
 	drawRect(&rect, color);
 
 	--rect.w;
 	--rect.h;
-	color = static_cast<Uint8>(static_cast<unsigned>(_color) + 4u);
+	color = static_cast<Uint8>(_color + 4u);
 
 	drawRect(&rect, color);
 
 	setPixelColor(
 			0,0,
-			static_cast<Uint8>(static_cast<unsigned>(_color) + 1u));
+			static_cast<Uint8>(_color + 1u));
 	setPixelColor(
 			0,
 			getHeight() - 1,
-			static_cast<Uint8>(static_cast<unsigned>(_color) + 4u));
+			static_cast<Uint8>(_color + 4u));
 	setPixelColor(
 			getWidth() - 1,
 			0,
-			static_cast<Uint8>(static_cast<unsigned>(_color) + 4u));
+			static_cast<Uint8>(_color + 4u));
 
-	color = static_cast<Uint8>(static_cast<unsigned>(_color) + 1u);
+	color = static_cast<Uint8>(_color + 1u);
 
 	switch (_shape)
 	{
@@ -176,7 +176,7 @@ void ArrowButton::draw()
 			for (
 					;
 					rect.w > 1u;
-					rect.w = static_cast<Uint16>(static_cast<unsigned>(rect.w) - 2u))
+					rect.w = static_cast<Uint16>(rect.w - 2u))
 			{
 				drawRect(&rect, color);
 				++rect.x;
@@ -201,7 +201,7 @@ void ArrowButton::draw()
 			for (
 					;
 					rect.w > 1u;
-					rect.w = static_cast<Uint16>(static_cast<unsigned>(rect.w) - 2u))
+					rect.w = static_cast<Uint16>(rect.w - 2u))
 			{
 				drawRect(&rect, color);
 				++rect.x;
@@ -219,13 +219,13 @@ void ArrowButton::draw()
 			for (
 					;
 					rect.w > 1u;
-					rect.w = static_cast<Uint16>(static_cast<unsigned>(rect.w) - 2u))
+					rect.w = static_cast<Uint16>(rect.w - 2u))
 			{
-				drawRect(&rect, static_cast<Uint8>(static_cast<unsigned>(color) + 2u));
+				drawRect(&rect, static_cast<Uint8>(color + 2u));
 				++rect.x;
 				--rect.y;
 			}
-			drawRect(&rect, static_cast<Uint8>(static_cast<unsigned>(color) + 2u));
+			drawRect(&rect, static_cast<Uint8>(color + 2u));
 
 			rect.x = 2; // draw arrow triangle 2
 			rect.y = 5;
@@ -235,7 +235,7 @@ void ArrowButton::draw()
 			for (
 					;
 					rect.w > 1u;
-					rect.w = static_cast<Uint16>(static_cast<unsigned>(rect.w) - 2u))
+					rect.w = static_cast<Uint16>(rect.w - 2u))
 			{
 				drawRect(&rect, color);
 				++rect.x;
@@ -253,13 +253,13 @@ void ArrowButton::draw()
 			for (
 					;
 					rect.w > 1u;
-					rect.w = static_cast<Uint16>(static_cast<unsigned>(rect.w) - 2u))
+					rect.w = static_cast<Uint16>(rect.w - 2u))
 			{
-				drawRect(&rect, static_cast<Uint8>(static_cast<unsigned>(color) + 2u));
+				drawRect(&rect, static_cast<Uint8>(color + 2u));
 				++rect.x;
 				++rect.y;
 			}
-			drawRect(&rect, static_cast<Uint8>(static_cast<unsigned>(color) + 2u));
+			drawRect(&rect, static_cast<Uint8>(color + 2u));
 
 			rect.x = // draw arrow triangle 2
 			rect.y = 2;
@@ -269,7 +269,7 @@ void ArrowButton::draw()
 			for (
 					;
 					rect.w > 1u;
-					rect.w = static_cast<Uint16>(static_cast<unsigned>(rect.w) - 2u))
+					rect.w = static_cast<Uint16>(rect.w - 2u))
 			{
 				drawRect(&rect, color);
 				++rect.x;
@@ -287,14 +287,14 @@ void ArrowButton::draw()
 			for (
 					;
 					rect.h < 5u;
-					rect.h = static_cast<Uint16>(static_cast<unsigned>(rect.h) + 2u))
+					rect.h = static_cast<Uint16>(rect.h + 2u))
 			{
-				drawRect(&rect, static_cast<Uint8>(static_cast<unsigned>(color) + 2u));
-				rect.x = static_cast<Sint16>(static_cast<int>(rect.x) + 2);
+				drawRect(&rect, static_cast<Uint8>(color + 2u));
+				rect.x = static_cast<Sint16>(rect.x + 2);
 				--rect.y;
 			}
 			rect.w = 1u;
-			drawRect(&rect, static_cast<Uint8>(static_cast<unsigned>(color) + 2u));
+			drawRect(&rect, static_cast<Uint8>(color + 2u));
 
 			rect.x = 3; // draw arrow triangle 2
 			rect.y = 4;
@@ -304,10 +304,10 @@ void ArrowButton::draw()
 			for (
 					;
 					rect.h < 5u;
-					rect.h = static_cast<Uint16>(static_cast<unsigned>(rect.h) + 2u))
+					rect.h = static_cast<Uint16>(rect.h + 2u))
 			{
 				drawRect(&rect, color);
-				rect.x = static_cast<Sint16>(static_cast<int>(rect.x) + 2);
+				rect.x = static_cast<Sint16>(rect.x + 2);
 				--rect.y;
 			}
 			rect.w = 1u;
@@ -323,15 +323,15 @@ void ArrowButton::draw()
 			for (
 					;
 					rect.h < 5u;
-					rect.h = static_cast<Uint16>(static_cast<unsigned>(rect.h) + 2u))
+					rect.h = static_cast<Uint16>(rect.h + 2u))
 			{
-				drawRect(&rect, static_cast<Uint8>(static_cast<unsigned>(color) + 2u));
-				rect.x = static_cast<Sint16>(static_cast<int>(rect.x) - 2);
+				drawRect(&rect, static_cast<Uint8>(color + 2u));
+				rect.x = static_cast<Sint16>(rect.x - 2);
 				--rect.y;
 			}
 			++rect.x;
 			rect.w = 1u;
-			drawRect(&rect, static_cast<Uint8>(static_cast<unsigned>(color) + 2u));
+			drawRect(&rect, static_cast<Uint8>(color + 2u));
 
 			rect.x = 6; // draw arrow triangle 2
 			rect.y = 4;
@@ -341,10 +341,10 @@ void ArrowButton::draw()
 			for (
 					;
 					rect.h < 5u;
-					rect.h = static_cast<Uint16>(static_cast<unsigned>(rect.h) + 2u))
+					rect.h = static_cast<Uint16>(rect.h + 2u))
 			{
 				drawRect(&rect, color);
-				rect.x = static_cast<Sint16>(static_cast<int>(rect.x) - 2);
+				rect.x = static_cast<Sint16>(rect.x - 2);
 				--rect.y;
 			}
 			++rect.x;
