+++ src/Interface/Text.cpp
@@ -658,8 +658,8 @@ void Text::draw()
 
 		++rect.x;
 		++rect.y;
-		rect.w = static_cast<Uint16>(static_cast<unsigned>(rect.w) - 2u);
-		rect.h = static_cast<Uint16>(static_cast<unsigned>(rect.h) - 2u);
+		rect.w = static_cast<Uint16>(rect.w - 2u);
+		rect.h = static_cast<Uint16>(rect.h - 2u);
 		this->drawRect(&rect, 0u);
 	}
 
@@ -690,8 +690,7 @@ void Text::draw()
 			y = 0;
 			break;
 		case ALIGN_MIDDLE:
-			y = static_cast<int>(std::ceil(
-				static_cast<double>(getHeight() - height) / 2.));
+			y = static_cast<int>(std::ceil((getHeight() - height) / 2.));
 			break;
 		case ALIGN_BOTTOM:
 			y = getHeight() - height;
